requirements_hybrid.txt
# Core LLM and AI libraries
langchain>=0.0.350
transformers>=4.36.0
torch>=2.0.0
google-generativeai>=0.3.0

# Data processing and storage
chromadb>=0.4.0
pandas>=1.5.0
numpy>=1.24.0

# Web interface
streamlit>=1.28.0

# Security and key management
keyring>=24.0.0
cryptography>=41.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Optional: GPU acceleration (install separately if needed)
# bitsandbytes>=0.41.0  # For 4-bit quantization
# accelerate>=0.24.0    # For model loading optimization

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Logging and monitoring
colorlog>=6.7.0

# File handling
PyYAML>=6.0
